refactor: refactored to more modern features
